Revert "Add X11DisplayManager to get display information"
This reverts commit a78be027cbbda783ab7dd94f1124289453445e71.
