+++ ui/ozone/platform/x11/x11_native_display_delegate.cc
@@ -16,8 +16,26 @@ X11NativeDisplayDelegate::X11NativeDisplayDelegate() = default;
 X11NativeDisplayDelegate::~X11NativeDisplayDelegate() = default;
 
 void X11NativeDisplayDelegate::Initialize() {
-  display_manager_ = std::make_unique<X11DisplayManagerOzone>();
-  display_manager_->SetObserver(this);
+  // This shouldn't be called twice.
+  DCHECK(!current_snapshot_);
+
+  XDisplay* display = gfx::GetXDisplay();
+  Screen* screen = DefaultScreenOfDisplay(display);
+  current_snapshot_.reset(new display::DisplaySnapshot(
+      XScreenNumberOfScreen(screen), gfx::Point(0, 0),
+      gfx::Size(WidthOfScreen(screen), HeightOfScreen(screen)),
+      display::DisplayConnectionType::DISPLAY_CONNECTION_TYPE_NONE, false,
+      false, false, gfx::ColorSpace(), "", base::FilePath(),
+      display::DisplaySnapshot::DisplayModeList(), std::vector<uint8_t>(),
+      nullptr, nullptr, 0, 0, gfx::Size()));
+  const int default_refresh = 60;
+  current_mode_.reset(new display::DisplayMode(
+      gfx::Size(WidthOfScreen(screen), HeightOfScreen(screen)), false,
+      default_refresh));
+  current_snapshot_->set_current_mode(current_mode_.get());
+
+  for (display::NativeDisplayObserver& observer : observers_)
+    observer.OnConfigurationChanged();
 }
 
 void X11NativeDisplayDelegate::TakeDisplayControl(
@@ -34,20 +52,16 @@ void X11NativeDisplayDelegate::GetDisplays(
     display::GetDisplaysCallback callback) {
   // TODO(msisov): Add support for multiple displays and dynamic display data
   // change.
-  if (displays_ready_)
-    display_manager_->GetDisplaysSnapshot(std::move(callback));
+  std::vector<display::DisplaySnapshot*> snapshot;
+  snapshot.push_back(current_snapshot_.get());
+  std::move(callback).Run(snapshot);
 }
 
 void X11NativeDisplayDelegate::Configure(const display::DisplaySnapshot& output,
                                          const display::DisplayMode* mode,
                                          const gfx::Point& origin,
                                          display::ConfigureCallback callback) {
-  NOTIMPLEMENTED();
-
-  // It should call |callback| after configuration.
-  // Even if we don't have implementation, it calls |callback| to finish the
-  // logic. otherwise, several tests from views_mus_unittests don't work.
-  std::move(callback).Run(true);
+  NOTREACHED();
 }
 
 void X11NativeDisplayDelegate::GetHDCPState(
@@ -92,11 +106,4 @@ X11NativeDisplayDelegate::GetFakeDisplayController() {
   return nullptr;
 }
 
-void X11NativeDisplayDelegate::OnOutputReadyForUse() {
-  if (!displays_ready_)
-    displays_ready_ = true;
-
-  for (display::NativeDisplayObserver& observer : observers_)
-    observer.OnConfigurationChanged();
-}
 }  // namespace ui
